requirements.txt
numpy~=1.22.4
opencv-python~=4.5.5.62
Shapely~=1.8.2
Pillow~=9.0.0